requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
chromadb>=0.5.0
openai>=1.50.0
beautifulsoup4>=4.12.0
httpx>=0.27.0
pyyaml>=6.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
